+++ src/main/java/by/mironenko/marketplace/service/impl/GameServiceImpl.java
@@ -6,10 +6,12 @@ import by.mironenko.marketplace.entity.Game;
 import by.mironenko.marketplace.exceptions.DaoException;
 import by.mironenko.marketplace.exceptions.ServiceException;
 import by.mironenko.marketplace.service.GameService;
+import by.mironenko.marketplace.service.Service;
 import lombok.RequiredArgsConstructor;
 import org.apache.log4j.Logger;
 
 import java.util.List;
+
 @RequiredArgsConstructor
 public class GameServiceImpl implements GameService {
     private static final Logger log = Logger.getLogger(GameServiceImpl.class);
@@ -19,6 +21,9 @@ public class GameServiceImpl implements GameService {
     @Override
     public void create(final Game game) throws ServiceException {
         log.info("<-SERVICE-> Creating new game...");
+        if (game == null || game.getPrice() <= 0.0) {
+            throw new ServiceException("Incorrect input parameters for creating game.");
+        }
         try {
             final GameDao gameDao = factory.getGameDao();
             gameDao.create(game);
@@ -41,6 +46,9 @@ public class GameServiceImpl implements GameService {
     @Override
     public Game findById(final Long id) throws ServiceException {
         log.info("<-SERVICE-> Finding game by ID...");
+        if (id <= 0) {
+            throw new ServiceException("Incorrect ID for buying game by ID.");
+        }
         try {
             final GameDao gameDao = factory.getGameDao();
             return gameDao.findById(id);
@@ -52,6 +60,9 @@ public class GameServiceImpl implements GameService {
     @Override
     public void update(final Game game) throws ServiceException {
         log.info("<-SERVICE-> Updating the game...");
+        if (game == null) {
+            throw new ServiceException("Cant define game for updating.");
+        }
         try {
             final GameDao gameDao = factory.getGameDao();
             gameDao.update(game);
@@ -63,6 +74,9 @@ public class GameServiceImpl implements GameService {
     @Override
     public void delete(final Long id) throws ServiceException {
         log.info("<-SERVICE-> Deleting the game...");
+        if (id <= 0) {
+            throw new ServiceException("Incorrect ID for buying game by ID.");
+        }
         try {
             final GameDao gameDao = factory.getGameDao();
             gameDao.delete(id);
@@ -74,6 +88,9 @@ public class GameServiceImpl implements GameService {
     @Override
     public Game findByGameName(final String gameName) throws ServiceException {
         log.info("<-SERVICE-> Finding game by name...");
+        if (gameName == null) {
+            throw new ServiceException("Incorrect game name.");
+        }
         try {
             final GameDao gameDao = factory.getGameDao();
             return gameDao.findByName(gameName);
@@ -85,6 +102,9 @@ public class GameServiceImpl implements GameService {
     @Override
     public List<Game> findByDeveloperName(final String developerName) throws ServiceException {
         log.info("<-SERVICE-> Finding games by developer name...");
+        if (developerName == null) {
+            throw new ServiceException("Incorrect developer name.");
+        }
         try {
             final GameDao gameDao = factory.getGameDao();
             return gameDao.findByDeveloper(developerName);
@@ -107,6 +127,9 @@ public class GameServiceImpl implements GameService {
     @Override
     public List<Game> findByPrice(final Double price) throws ServiceException {
         log.info("<-SERVICE-> Finding games by price...");
+        if (price <= 0.0) {
+            throw new ServiceException("Incorrect price for finding game by price.");
+        }
         try {
             final GameDao gameDao = factory.getGameDao();
             return gameDao.findByPrice(price);
